error handling refactoring

+++ api/controllers/groupController.js
@@ -1,10 +1,11 @@
 import { groupService } from '../services/groupService';
+const createError = require('http-errors');
 
 const create = async (req, res, next) => {
     try {
         const group = await groupService.getGroupByName(req.body.name);
         if (group) {
-            res.status(409).send(`Group name ${req.body.name} already exists!`);
+            throw createError.Conflict(`Group name ${req.body.login} already exists!`);
         } else {
             const groupData = await groupService.createGroup(req.body);
             res.status(201).json(groupData.id);
@@ -29,7 +30,7 @@ const getById = async (req, res, next) => {
         if (group) {
             res.json(group);
         } else {
-            res.status(404).send(`Group with id ${req.params.id} not found!`);
+            throw createError.NotFound(`Group id ${req.params.id} not found!`);
         }
     } catch (error) {
         return next(error);
@@ -39,12 +40,12 @@ const getById = async (req, res, next) => {
 const update = async (req, res, next) => {
     try {
         const group = await groupService.getGroupById(req.params.id);
-        if (!group) {
-            res.status(404).send(`Group with id ${req.params.id} not found!`);
-        } else {
+        if (group) {
             await groupService.updateGroup(req.params.id, req.body);
             await group.reload();
             res.json(group);
+        } else {
+            throw createError.NotFound(`Group id ${req.params.id} not found!`);
         }
     } catch (error) {
         return next(error);
@@ -53,8 +54,13 @@ const update = async (req, res, next) => {
 
 const remove = async (req, res, next) => {
     try {
-        await groupService.deleteGroupById(req.params.id);
-        res.status(200).end();
+        const group = await groupService.getGroupById(req.params.id);
+        if (group) {
+            await groupService.deleteGroupById(req.params.id);
+            res.status(200).end();
+        } else {
+            throw createError.NotFound(`Group id ${req.params.id} not found!`);
+        }
     } catch (error) {
         return next(error);
     }
@@ -63,17 +69,16 @@ const remove = async (req, res, next) => {
 const addUsers = async (req, res, next) => {
     try {
         const group = await groupService.getGroupById(req.params.id);
-        if (!group) {
-            res.status(404).send(`Group with id ${req.params.id} not found!`);
-        } else {
+        if (group) {
             const groupUpdated = await groupService.addUsersToGroup(req.params.id, req.body.usersIds);
             res.json(groupUpdated);
+        } else {
+            throw createError.NotFound(`Group id ${req.params.id} not found!`);
         }
     } catch (error) {
         return next(error);
     }
 };
-
 
 export {
     addUsers,
